refactor email hint spec
